+++ CHANGES.md
@@ -45,6 +45,7 @@ Changes in 12.5
 - JS remove inline code from program in Courses.php & Courses.js
 - JS remove inline code from program in MassDrops.php & MassDrops.js
 - JS remove inline code from program in MassRequests.php & MassRequests.js
+- JS remove inline code from program in MassSchedule.php & MassSchedule.js
 
 Changes in 12.4.3
 -----------------